Refactored frontend automations components
