+++ apps/posts/src/views/Automations/components/canvas/step-picker.tsx
